requirements.txt
Flask
Pillow
pdf2image
python-dotenv
requests
